split powerups.py into submodules for matplotlib and plotly
