+++ pymatviz/powerups/both.py
@@ -1,6 +1,4 @@
-"""Powerups/enhancements such as parity lines, annotations and marginals for matplotlib
-and plotly figures.
-"""
+"""Powerups that can be applied to both matplotlib and plotly figures."""
 
 from __future__ import annotations
 
@@ -8,7 +6,6 @@ from typing import TYPE_CHECKING, Any
 
 import matplotlib.pyplot as plt
 import numpy as np
-import plotly.express as px
 import plotly.graph_objects as go
 import sklearn
 from sklearn.metrics import mean_absolute_percentage_error as mape
@@ -32,129 +29,8 @@ from pymatviz.utils import (
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-    from matplotlib.gridspec import GridSpec
     from matplotlib.offsetbox import AnchoredText
-    from matplotlib.text import Annotation
     from numpy.typing import ArrayLike
-    from plotly.basedatatypes import BaseTraceType
-
-
-def with_marginal_hist(
-    xs: ArrayLike,
-    ys: ArrayLike,
-    cell: GridSpec | None = None,
-    bins: int = 100,
-    fig: plt.Figure | plt.Axes | None = None,
-) -> plt.Axes:
-    """Call before creating a plot and use the returned `ax_main` for all
-    subsequent plotting ops to create a grid of plots with the main plot in the
-    lower left and narrow histograms along its x- and/or y-axes displayed above
-    and near the right edge.
-
-    Args:
-        xs (array): Marginal histogram values along x-axis.
-        ys (array): Marginal histogram values along y-axis.
-        cell (GridSpec, optional): Cell of a plt GridSpec at which to add the
-            grid of plots. Defaults to None.
-        bins (int, optional): Resolution/bin count of the histograms. Defaults to 100.
-        fig (Figure, optional): matplotlib Figure or Axes to add the marginal histograms
-            to. Defaults to None.
-
-    Returns:
-        plt.Axes: The matplotlib Axes to be used for the main plot.
-    """
-    if fig is None or isinstance(fig, plt.Axes):
-        ax_main = fig or plt.gca()
-        fig = ax_main.figure
-
-    gs = (cell.subgridspec if cell else fig.add_gridspec)(
-        2, 2, width_ratios=(6, 1), height_ratios=(1, 5), wspace=0, hspace=0
-    )
-
-    ax_main = fig.add_subplot(gs[1, 0])
-    ax_histx = fig.add_subplot(gs[0, 0], sharex=ax_main)
-    ax_histy = fig.add_subplot(gs[1, 1], sharey=ax_main)
-
-    # x_hist
-    ax_histx.hist(xs, bins=bins, rwidth=0.8)
-    ax_histx.axis("off")
-
-    # y_hist
-    ax_histy.hist(ys, bins=bins, rwidth=0.8, orientation="horizontal")
-    ax_histy.axis("off")
-
-    return ax_main
-
-
-def annotate_bars(
-    ax: plt.Axes | None = None,
-    *,
-    v_offset: float = 10,
-    h_offset: float = 0,
-    labels: Sequence[str | int | float] | None = None,
-    fontsize: int = 14,
-    y_max_headroom: float = 1.2,
-    adjust_test_pos: bool = False,
-    **kwargs: Any,
-) -> None:
-    """Annotate each bar in bar plot with a label.
-
-    Args:
-        ax (Axes): The matplotlib axes to annotate.
-        v_offset (int): Vertical offset between the labels and the bars.
-        h_offset (int): Horizontal offset between the labels and the bars.
-        labels (list[str]): Labels used for annotating bars. If not provided, defaults
-            to the y-value of each bar.
-        fontsize (int): Annotated text size in pts. Defaults to 14.
-        y_max_headroom (float): Will be multiplied with the y-value of the tallest bar
-            to increase the y-max of the plot, thereby making room for text above all
-            bars. Defaults to 1.2.
-        adjust_test_pos (bool): If True, use adjustText to prevent overlapping labels.
-            Defaults to False.
-        **kwargs: Additional arguments (rotation, arrowprops, etc.) are passed to
-            ax.annotate().
-    """
-    ax = ax or plt.gca()
-
-    if labels is None:
-        labels = [int(patch.get_height()) for patch in ax.patches]
-    elif len(labels) != len(ax.patches):
-        raise ValueError(
-            f"Got {len(labels)} labels but {len(ax.patches)} bars to annotate"
-        )
-
-    y_max: float = 0
-    texts: list[Annotation] = []
-    for rect, label in zip(ax.patches, labels):
-        y_pos = rect.get_height()
-        x_pos = rect.get_x() + rect.get_width() / 2 + h_offset
-
-        if ax.get_yscale() == "log":
-            y_pos += np.log(max(1, v_offset))
-        else:
-            y_pos += v_offset
-
-        y_max = max(y_max, y_pos)
-
-        txt = f"{label:,}" if isinstance(label, (int, float)) else label
-        # place label at end of the bar and center horizontally
-        anno = ax.annotate(
-            txt, (x_pos, y_pos), ha="center", fontsize=fontsize, **kwargs
-        )
-        texts.append(anno)
-
-    # ensure enough vertical space to display label above highest bar
-    ax.set(ylim=(None, y_max * y_max_headroom))
-    if adjust_test_pos:
-        try:
-            from adjustText import adjust_text
-
-            adjust_text(texts, ax=ax)
-        except ImportError as exc:
-            raise ImportError(
-                "adjustText not installed, falling back to default matplotlib "
-                "label placement. Use pip install adjustText."
-            ) from exc
 
 
 @validate_fig
@@ -422,142 +298,3 @@ def add_best_fit_line(
         return fig
 
     raise ValueError(f"Unsupported {backend=}. Must be one of {VALID_BACKENDS}")
-
-
-def add_ecdf_line(
-    fig: go.Figure,
-    values: ArrayLike = (),
-    trace_idx: int = 0,
-    trace_kwargs: dict[str, Any] | None = None,
-    **kwargs: Any,
-) -> go.Figure:
-    """Add an empirical cumulative distribution function (ECDF) line to a plotly figure.
-
-    Support for matplotlib planned but not implemented. PRs welcome.
-
-    Args:
-        fig (go.Figure): plotly figure to add the ECDF line to.
-        values (array, optional): Values to compute the ECDF from. Defaults to () which
-            means use the x-values of trace at trace_idx in fig.
-        trace_idx (int, optional): Index of the trace whose x-values to use for
-            computing the ECDF. Defaults to 0. Unused if values is not empty.
-        trace_kwargs (dict[str, Any], optional): Passed to trace_ecdf.update().
-            Defaults to None. Use e.g. to set trace name (default "Cumulative") or
-            line_color (default "gray").
-        **kwargs: Passed to fig.add_trace().
-
-    Returns:
-        Figure: Figure with added ECDF line.
-    """
-    trace_kwargs = trace_kwargs or {}
-    valid_fig_types = (go.Figure,)
-    if not isinstance(fig, valid_fig_types):
-        type_names = " | ".join(
-            f"{t.__module__}.{t.__qualname__}" for t in valid_fig_types
-        )
-        raise TypeError(f"{fig=} must be instance of {type_names}")
-
-    target_trace: BaseTraceType = fig.data[trace_idx]
-    if values is None or len(values) == 0:
-        if isinstance(target_trace, (go.Histogram, go.Scatter, go.Scattergl)):
-            values = target_trace.x
-        elif isinstance(target_trace, go.Bar):
-            xs, ys = target_trace.x, target_trace.y
-            # if xs are bin edges, drop last and interpret as bin centers
-            if len(xs) + 1 == len(ys):
-                xs = xs[:-1]
-            values = np.repeat(xs, ys)
-        else:
-            cls = type(target_trace)
-            qual_name = f"{cls.__module__}.{cls.__qualname__}"
-            raise ValueError(
-                f"Cannot auto-determine x-values for ECDF from {qual_name}, "
-                "pass values explicitly. Currently only Histogram, Scatter, Box, "
-                "and Violin traces are supported and may well need more testing. "
-                "Please report issues at https://github.com/janosh/pymatviz/issues."
-            )
-
-    ecdf_trace = px.ecdf(values).data[0]
-
-    # if fig has facets, add ECDF to all subplots
-    fig_add_trace_defaults = {} if fig._grid_ref is None else dict(row="all", col="all")  # noqa: SLF001
-    fig.add_trace(ecdf_trace, **fig_add_trace_defaults | kwargs)
-
-    # get xlabel falling back on 'x' if not set
-    xlabel = fig.layout.xaxis.title.text or "x"
-    tooltip_template = f"{xlabel}: %{{x}}<br>Percent: %{{y:.2%}}<extra></extra>"
-    fig.data[-1].hovertemplate = tooltip_template
-
-    # move ECDF line to secondary y-axis
-    # set color to darkened version of primary y-axis color
-    name = uniq_name = trace_kwargs.get("name", "Cumulative")
-    trace_names = [trace.name for trace in fig.data]
-    name_suffix = 1
-    while uniq_name in trace_names:
-        name_suffix += 1
-        uniq_name = f"{name} {name_suffix}"
-
-    if "marker" in target_trace and target_trace.marker["color"] is None:
-        dev_trace = fig.full_figure_for_development(warn=False).data[trace_idx]
-        # NOTE unsure if this has any downstream effects, should just be applying
-        # the color that plotly would have later assigned to the trace anyway
-        target_trace.marker = dev_trace.marker
-
-    # set default legendgroup of target trace if not already set
-    legendgroup = target_trace.legendgroup or target_trace.name
-    target_trace.legendgroup = legendgroup
-
-    # set line color to be the same as the target trace's marker color
-    target_color = target_trace.marker.color if "marker" in target_trace else "gray"
-    trace_defaults = dict(
-        yaxis="y2",
-        name=uniq_name,
-        line=dict(
-            color=target_color,
-            dash="solid",
-        ),
-        legendgroup=legendgroup,
-    )
-    trace_kwargs = trace_defaults | trace_kwargs
-    fig.data[-1].update(**trace_kwargs)
-
-    # line_color becomes target_color via trace_defaults if a different color was not
-    # already set in trace_kwargs
-    line_color = trace_kwargs.get(
-        "line_color", trace_kwargs.get("line", {}).get("color")
-    )
-
-    yaxis_defaults = dict(
-        title=uniq_name,
-        side="right",
-        overlaying="y",
-        range=(0, 1),
-        showgrid=False,
-        color=line_color,
-    )
-    # make secondary ECDF y-axis inherit primary y-axis styles
-    yaxis2_layout = getattr(fig.layout, "yaxis2", {})
-    if yaxis2_layout:
-        # convert to dict
-        yaxis2_layout = yaxis2_layout._props  # type: ignore[union-attr] # noqa: SLF001
-    fig.layout.yaxis2 = yaxis_defaults | yaxis2_layout
-
-    return fig
-
-
-# buttons to toggle log/linear y-axis. apply to a plotly figure like this:
-# fig.layout.updatemenus = [toggle_log_linear_y_axis]
-# use toggle_log_linear_y_axis | dict(x=1, y=0, ...) to customize
-toggle_log_linear_y_axis = dict(
-    type="buttons",
-    direction="left",
-    buttons=[
-        dict(args=[{"yaxis.type": "linear"}], label="Linear Y", method="relayout"),
-        dict(args=[{"yaxis.type": "log"}], label="Log Y", method="relayout"),
-    ],
-    pad={"r": 10, "t": 10},
-    showactive=True,
-    x=0.1,
-    y=1,
-    yanchor="top",
-)
